requirements.txt
openai>=1.0.0
praw>=7.8.0
PyGithub>=2.6.0
psutil>=7.0.0
pyserial>=3.5
watchdog>=3.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
